requirements.txt
flask
openai==1.51.2
python-dotenv
